requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==40.0.1
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.103.2
h11==0.14.0
idna==3.4
motor==3.1.1
passlib==1.7.4
Pillow==10.0.1
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn[standard]==0.23.2